create seperate class for calling graphql API
